1 add a screeninfo and width arg to the inRect() functions 2 implement line, ray, segment and circle inRect() functions 3 fix gcc 3 errors and warnings
svn path=/branches/make_kdeedu_cool/kdeedu/kig/; revision=214816

+++ misc/screeninfo.h
@@ -44,7 +44,7 @@ public:
 
   double pixelWidth() const;
 
-  double normalMiss( double width ) const;
+  double normalMiss( int width ) const;
 
   const Rect& shownRect() const;
